chore: address reviews
