+++ frontend/packages/app/src/components/timesheet-row/components/inline-time-entry/index.tsx
@@ -4,6 +4,7 @@
 import { useCallback, useMemo, useRef, useState } from "react";
 import { Accordion } from "@base-ui/react/accordion";
 import { floatToTime, mergeClassNames as cn } from "@next-pms/design-system";
+import { stripTags } from "@next-pms/design-system/utils";
 import { Badge, Button, TextEditor, useToasts } from "@rtcamp/frappe-ui-react";
 import { FrappeError, useFrappePostCall } from "frappe-react-sdk";
 import { Edit, Pen, Plus, Trash2 } from "lucide-react";
@@ -11,7 +12,7 @@ import { Edit, Pen, Plus, Trash2 } from "lucide-react";
 /**
  * Internal Dependencies
  */
-import { parseFrappeErrorMsg, stripTags } from "@/lib/utils";
+import { parseFrappeErrorMsg } from "@/lib/utils";
 import { TaskDataItemProps } from "@/types/timesheet";
 import { useInlineTimeEntryForm } from "./form";
 import { TimeEntryForm } from "./timeEntryForm";
